Move more into CircuitInterceptingNavigator
