+++ circuitx/navigation/src/androidMain/kotlin/com/slack/circuitx/navigation/intercepting/CircuitInterceptingNavigator.android.kt
