+++ circuitx/navigation/src/jvmMain/kotlin/com/slack/circuitx/navigation/intercepting/CircuitInterceptingNavigator.jvm.kt
