+++ circuitx/navigation/src/nativeMain/kotlin/com/slack/circuitx/navigation/intercepting/CircuitInterceptingNavigator.native.kt
